+++ stellargraph/core/graph.py
@@ -203,8 +203,6 @@ class StellarGraph:
             Deprecated, use :meth:`from_networkx`.
         edge_type_name:
             Deprecated, use :meth:`from_networkx`.
-        edge_weight_label:
-            Deprecated, use :meth:`from_networkx`.
         node_features:
             Deprecated, use :meth:`from_networkx`.
     """
@@ -225,19 +223,9 @@ class StellarGraph:
         graph=None,
         node_type_name=globalvar.TYPE_ATTR_NAME,
         edge_type_name=globalvar.TYPE_ATTR_NAME,
-        edge_weight_label=None,
         node_features=None,
     ):
         import networkx
-
-        # support for legacy arguments, translate to the new form
-        if edge_weight_label is not None:
-            # `edge_weight_label` -> `edge_weight_column`
-            warnings.warn(
-                "the 'edge_weight_label' parameter has been replaced by 'edge_weight_column'",
-                DeprecationWarning,
-            )
-            edge_weight_column = edge_weight_label
 
         if isinstance(nodes, networkx.Graph):
             # `StellarGraph(nx_graph)` -> `graph`
@@ -297,9 +285,6 @@ class StellarGraph:
         edge_type_default=globalvar.EDGE_TYPE_DEFAULT,
         node_features=None,
         dtype="float32",
-        node_type_name=None,
-        edge_type_name=None,
-        edge_weight_label=None,
     ):
         """
         Construct a ``StellarGraph`` object from a NetworkX graph::
@@ -377,33 +362,10 @@ class StellarGraph:
             edge_weight_attr (str, optional):
                 The name of the attribute to use as the weight of edges.
 
-            node_type_name: Deprecated, use ``node_type_attr``.
-            edge_type_name: Deprecated, use ``edge_type_attr``.
-            edge_weight_label: Deprecated, use ``edge_weight_attr``.
-
         Returns:
             A ``StellarGraph`` (if ``graph`` is undirected) or ``StellarDiGraph`` (if ``graph`` is
             directed) instance representing the data in ``graph`` and ``node_features``.
         """
-        if node_type_name is not None:
-            warnings.warn(
-                "the 'node_type_name' parameter has been replaced by 'node_type_attr'",
-                DeprecationWarning,
-            )
-            node_type_attr = node_type_name
-        if edge_type_name is not None:
-            warnings.warn(
-                "the 'edge_type_name' parameter has been replaced by 'edge_type_attr'",
-                DeprecationWarning,
-            )
-            edge_type_attr = edge_type_name
-        if edge_weight_label is not None:
-            warnings.warn(
-                "the 'edge_weight_label' parameter has been replaced by 'edge_weight_attr'",
-                DeprecationWarning,
-            )
-            edge_weight_attr = edge_weight_label
-
         nodes, edges = convert.from_networkx(
             graph,
             node_type_attr=node_type_attr,
@@ -685,14 +647,12 @@ class StellarGraph:
         Returns:
             A dictionary of node type and integer feature size.
         """
-        all_sizes = self._nodes.feature_sizes()
-        if node_types is None:
-            return all_sizes
+        all_sizes = self._nodes.feature_info()
 
-        node_types = set(node_types)
-        return {
-            type_name: size for type_name, size in all_sizes if type_name in node_types
-        }
+        if node_types is None:
+            node_types = all_sizes.keys()
+
+        return {type_name: all_sizes[type_name][0] for type_name in node_types}
 
     def check_graph_for_ml(self, features=True):
         """
@@ -851,9 +811,18 @@ class StellarGraph:
         lines.append("")
         lines.append(" Node types:")
 
+        feature_info = self._nodes.feature_info()
         for nt in gs.node_types:
             nodes = self.nodes_of_type(nt)
             lines.append(f"  {nt}: [{len(nodes)}]")
+
+            feature_size, feature_dtype = feature_info[nt]
+            if feature_size > 0:
+                feature_text = f"{feature_dtype.name} vector, length {feature_size}"
+            else:
+                feature_text = "none"
+            lines.append(f"    Features: {feature_text}")
+
             edge_types = ", ".join(str_edge_type(et) for et in gs.schema[nt])
             lines.append(f"    Edge types: {edge_types}")
 
@@ -871,13 +840,7 @@ class StellarGraph:
 
     def create_graph_schema(self, nodes=None):
         """
-        Create graph schema in dict of dict format from current graph.
-
-        Note the assumption we make that there is only one
-        edge of a particular edge type per node pair.
-
-        This means that specifying an edge by node0, node1 and edge type
-        is unique.
+        Create graph schema from the current graph.
 
         Arguments:
             nodes (list): A list of node IDs to use to build schema. This must
@@ -988,31 +951,140 @@ class StellarGraph:
         adj.sum_duplicates()
         return adj
 
+    def subgraph(self, nodes):
+        """
+        Compute the node-induced subgraph implied by ``nodes``.
+
+        Args:
+            nodes (iterable): The nodes in the subgraph.
+
+        Returns:
+            A :class:`StellarGraph` or :class:`StellarDiGraph` instance containing only the nodes in
+            ``nodes``, and any edges between them in ``self``. It contains the same node & edge
+            types, node features and edge weights as in ``self``.
+        """
+
+        node_ilocs = self._nodes.ids.to_iloc(nodes, strict=True)
+        node_types = self._nodes.type_of_iloc(node_ilocs)
+        node_type_to_ilocs = pd.Series(node_ilocs, index=node_types).groupby(level=0)
+
+        node_frames = {
+            type_name: pd.DataFrame(
+                self._nodes.features(type_name, ilocs),
+                index=self._nodes.ids.from_iloc(ilocs),
+            )
+            for type_name, ilocs in node_type_to_ilocs
+        }
+
+        # FIXME(#985): this is O(edges in graph) but could potentially be optimised to O(edges in
+        # graph incident to `nodes`), which could be much fewer if `nodes` is small
+        edge_ilocs = np.where(
+            np.isin(self._edges.sources, nodes) & np.isin(self._edges.targets, nodes)
+        )
+        edge_frame = pd.DataFrame(
+            {
+                "id": self._edges.ids.from_iloc(edge_ilocs),
+                globalvar.SOURCE: self._edges.sources[edge_ilocs],
+                globalvar.TARGET: self._edges.targets[edge_ilocs],
+                globalvar.WEIGHT: self._edges.weights[edge_ilocs],
+            },
+            index=self._edges.type_of_iloc(edge_ilocs),
+        )
+        edge_frames = {
+            type_name: df.set_index("id")
+            for type_name, df in edge_frame.groupby(level=0)
+        }
+
+        cls = StellarDiGraph if self.is_directed() else StellarGraph
+        return cls(node_frames, edge_frames)
+
+    def connected_components(self):
+        """
+        Compute the connected components in this graph, ordered by size.
+
+        The nodes in the largest component can be computed with ``nodes =
+        next(graph.connected_components())``. The node IDs returned by this method can be used to
+        compute the corresponding subgraph with ``graph.subgraph(nodes)``.
+
+        For directed graphs, this computes the weakly connected components. This effectively
+        treating each edge as undirected.
+
+        Returns:
+            An iterator over sets of node IDs in each connected component, from the largest (most nodes)
+            to smallest (fewest nodes).
+        """
+
+        adj = self.to_adjacency_matrix()
+        count, cc_labels = sps.csgraph.connected_components(adj, directed=False)
+        cc_sizes = np.bincount(cc_labels, minlength=count)
+        cc_by_size = np.argsort(cc_sizes)[::-1]
+
+        return (
+            self._nodes.ids.from_iloc(cc_labels == cc_label) for cc_label in cc_by_size
+        )
+
     def to_networkx(
         self,
-        node_type_name=globalvar.TYPE_ATTR_NAME,
-        edge_type_name=globalvar.TYPE_ATTR_NAME,
-        edge_weight_label=globalvar.WEIGHT,
-        feature_name=globalvar.FEATURE_ATTR_NAME,
+        node_type_attr=globalvar.TYPE_ATTR_NAME,
+        edge_type_attr=globalvar.TYPE_ATTR_NAME,
+        edge_weight_attr=globalvar.WEIGHT,
+        feature_attr=globalvar.FEATURE_ATTR_NAME,
+        node_type_name=None,
+        edge_type_name=None,
+        edge_weight_label=None,
+        feature_name=None,
     ):
         """
         Create a NetworkX MultiGraph or MultiDiGraph instance representing this graph.
 
         Args:
-            node_type_name (str): the name of the attribute to use to store a node's type (or label).
+            node_type_attr (str): the name of the attribute to use to store a node's type (or label).
 
-            edge_type_name (str): the name of the attribute to use to store a edge's type (or label).
+            edge_type_attr (str): the name of the attribute to use to store a edge's type (or label).
 
-            edge_weight_label (str): the name of the attribute to use to store a edge's weight.
+            edge_weight_attr (str): the name of the attribute to use to store a edge's weight.
 
-            feature_name (str, optional): the name of the attribute to use to store a node's feature
+            feature_attr (str, optional): the name of the attribute to use to store a node's feature
                 vector; if ``None``, feature vectors are not stored within each node.
+
+            node_type_name (str): Deprecated, use ``node_type_attr``.
+            edge_type_name (str): Deprecated, use ``edge_type_attr``.
+            edge_weight_label (str): Deprecated, use ``edge_weight_attr``.
+            feature_name (str, optional): Deprecated, use ``feature_attr``.
 
         Returns:
              An instance of `networkx.MultiDiGraph` (if directed) or `networkx.MultiGraph` (if
              undirected) containing all the nodes & edges and their types & features in this graph.
         """
         import networkx
+
+        if node_type_name is not None:
+            warnings.warn(
+                "the 'node_type_name' parameter has been replaced by 'node_type_attr'",
+                DeprecationWarning,
+            )
+            node_type_attr = node_type_name
+
+        if edge_type_name is not None:
+            warnings.warn(
+                "the 'edge_type_name' parameter has been replaced by 'edge_type_attr'",
+                DeprecationWarning,
+            )
+            edge_type_attr = edge_type_name
+
+        if edge_weight_label is not None:
+            warnings.warn(
+                "the 'edge_weight_label' parameter has been replaced by 'edge_weight_attr'",
+                DeprecationWarning,
+            )
+            edge_weight_attr = edge_weight_label
+
+        if feature_name is not None:
+            warnings.warn(
+                "the 'feature_name' parameter has been replaced by 'feature_attr'",
+                DeprecationWarning,
+            )
+            feature_attr = feature_name
 
         if self.is_directed():
             graph = networkx.MultiDiGraph()
@@ -1021,14 +1093,14 @@ class StellarGraph:
 
         for ty in self.node_types:
             node_ids = self.nodes_of_type(ty)
-            ty_dict = {node_type_name: ty}
+            ty_dict = {node_type_attr: ty}
 
-            if feature_name is not None:
+            if feature_attr is not None:
                 features = self.node_features(node_ids, node_type=ty)
 
                 for node_id, node_features in zip(node_ids, features):
                     graph.add_node(
-                        node_id, **ty_dict, **{feature_name: node_features},
+                        node_id, **ty_dict, **{feature_attr: node_features},
                     )
             else:
                 graph.add_nodes_from(node_ids, **ty_dict)
@@ -1040,7 +1112,7 @@ class StellarGraph:
             self._edges.weights,
         )
         graph.add_edges_from(
-            (src, dst, {edge_type_name: type_, edge_weight_label: weight})
+            (src, dst, {edge_type_attr: type_, edge_weight_attr: weight})
             for src, dst, type_, weight in iterator
         )
 
@@ -1145,7 +1217,6 @@ class StellarDiGraph(StellarGraph):
         graph=None,
         node_type_name=globalvar.TYPE_ATTR_NAME,
         edge_type_name=globalvar.TYPE_ATTR_NAME,
-        edge_weight_label=None,
         node_features=None,
     ):
         super().__init__(
@@ -1162,6 +1233,5 @@ class StellarDiGraph(StellarGraph):
             graph=graph,
             node_type_name=node_type_name,
             edge_type_name=edge_type_name,
-            edge_weight_label=edge_weight_label,
             node_features=node_features,
         )